+++ src/explorerCommands/utils.ts
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
+import { isJavaIdentifier, isKeyword } from "../utility";
 import { DataNode } from "../views/dataNode";
 
 export function isMutable(node: DataNode): boolean {
@@ -10,4 +11,22 @@ export function isMutable(node: DataNode): boolean {
 
     const contextValue = node.computeContextValue();
     return packageExp.test(contextValue) || resourceOrTypeExp.test(contextValue);
+}
+
+export function checkJavaQualifiedName(value: string): string {
+    if (!value || !value.trim()) {
+        return "Input cannot be empty.";
+    }
+
+    for (const part of value.split(".")) {
+        if (isKeyword(part)) {
+            return `Keyword '${part}' cannot be used.`;
+        }
+
+        if (!isJavaIdentifier(part)) {
+            return `Invalid Java qualified name.`;
+        }
+    }
+
+    return "";
 }
